// src/app/models/articulo/articulo.model.ts
export class Articulo {
    cod_Barras?: string;
    descripcion_Corta?: string;
    id_Clasificacion?: string;
    articulo_Disponible?: string;
    last_Update_Inventory?: string;
    stock_Min?: string;
    cod_nterno?: string;
    descripcion?: string;
    precio_Venta?: string;
    stock_Max?: string;
    cantidad_Um?: string;
    precio_Compra?: string;
    utilidad?: string;
    cod_Asociado?: string;
    id_Unidad?: string;
    id_Proveedor?: string;
    tipo_Articulo?: string;
    stock?: string;
    kit?: string;
    iva?: string;
    activo?: string;
    kit_Fecha_Ini?: string;
    visible?: string;
    fecha_Registro?: string;
    kit_Fecha_Fin?: string;
    cve_Producto?: string;
    puntos?: string;
}